requirements.txt
  
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.1
python-dotenv>=1.0.0
requests>=2.31.0

# Development dependencies
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
